Move plugin management into new module and add sections
