+++ client/jetpack-cloud/sections/agency-dashboard/controller.tsx
@@ -1,6 +1,5 @@
 import DashboardOverview from './dashboard-overview';
 import Header from './header';
-import PluginManagement from './plugin-management';
 import DashboardSidebar from './sidebar';
 
 export function agencyDashboardContext( context: PageJS.Context, next: VoidFunction ): void {
@@ -15,12 +14,5 @@ export function agencyDashboardContext( context: PageJS.Context, next: VoidFunct
 	context.primary = (
 		<DashboardOverview search={ search } currentPage={ currentPage } filter={ filter } />
 	);
-	next();
-}
-
-export function pluginManagementContext( context: PageJS.Context, next: VoidFunction ): void {
-	context.header = <Header />;
-	context.secondary = <DashboardSidebar path={ context.path } />;
-	context.primary = <PluginManagement />;
 	next();
 }
