+++ client/jetpack-cloud/style.scss
@@ -300,7 +300,7 @@ html {
 	font-size: $font-body;
 }
 
-.sections__logo, .dashboard-overview__logo, .site-content__logo, .plugin-management__logo {
+.sections__logo, .dashboard-overview__logo, .site-content__logo, .plugins-overview__logo {
 	fill: var( --color-neutral-10 );
 	color: var( --color-neutral-10 );
 	position: fixed;
