+++ config/_shared.json
@@ -158,6 +158,7 @@
 	"sections": {
 		"jetpack-cloud": false,
 		"jetpack-cloud-agency-dashboard": false,
+		"jetpack-cloud-plugin-management": false,
 		"jetpack-cloud-agency-signup": false,
 		"jetpack-cloud-auth": false,
 		"jetpack-cloud-partner-portal": false,
